move beamlattice to its own package
